Export Function
The function works for both active and passive now. It does error handling well in case path does not exist. If the file is named the same then scan data is appended to the same file. This will be refined much more over time. The output function is being repeated twice since my previous solution broke the code.